requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
openai==1.46.1
python-dotenv==1.0.1